# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "RelayOS Test Target API",
    "version": "0.1.0",
    "description": "Synthetic customer system for testing RelayOS tool connectivity. Not a production service."
  },
  "servers": [
    {
      "url": "http://localhost:3100",
      "description": "Local / docker-compose"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    { "bearerAuth": [] }
  ],
  "paths": {
    "/health": {
      "get": {
        "operationId": "getHealth",
        "summary": "Health check",
        "description": "Returns service health status. No authentication required.",
        "security": [],
        "responses": {
          "200": {
            "description": "Service is healthy",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["status"],
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "ok"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/streetlights": {
      "get": {
        "operationId": "getStreetlights",
        "summary": "List streetlights",
        "description": "Returns all streetlights. Optionally filter by status.",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "description": "Filter by streetlight status",
            "schema": {
              "type": "string",
              "enum": ["ok", "fault", "assigned"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Array of streetlights",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": ["id", "latitude", "longitude", "status", "last_seen"],
                    "properties": {
                      "id": { "type": "integer", "example": 1 },
                      "latitude": { "type": "number", "example": 30.26723 },
                      "longitude": { "type": "number", "example": -97.74312 },
                      "status": {
                        "type": "string",
                        "enum": ["ok", "fault", "assigned"],
                        "example": "fault"
                      },
                      "last_seen": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-15T08:30:00Z"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["error"],
                  "properties": {
                    "error": { "type": "string", "example": "Unauthorized" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/streetlights/{id}": {
      "patch": {
        "operationId": "patchStreetlight",
        "summary": "Update streetlight status",
        "description": "Updates the status of a single streetlight. Primarily used as a test-data reset affordance — e.g. to put an assigned light back to 'fault' without direct database access.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Streetlight ID",
            "schema": { "type": "integer", "example": 8 }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["status"],
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": ["ok", "fault", "assigned"],
                    "description": "New status for the streetlight",
                    "example": "fault"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated streetlight row",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["id", "latitude", "longitude", "status", "last_seen"],
                  "properties": {
                    "id": { "type": "integer", "example": 8 },
                    "latitude": { "type": "number", "example": 30.26601 },
                    "longitude": { "type": "number", "example": -97.73801 },
                    "status": {
                      "type": "string",
                      "enum": ["ok", "fault", "assigned"],
                      "example": "fault"
                    },
                    "last_seen": {
                      "type": "string",
                      "format": "date-time",
                      "example": "2024-01-15T08:30:00Z"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid status value",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["error"],
                  "properties": {
                    "error": { "type": "string", "example": "status must be one of: ok, fault, assigned" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["error"],
                  "properties": {
                    "error": { "type": "string", "example": "Unauthorized" }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Streetlight ID not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["error"],
                  "properties": {
                    "error": { "type": "string", "example": "Streetlight 99 not found" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/work-orders/{id}": {
      "get": {
        "operationId": "getWorkOrder",
        "summary": "Get one work order with its attached streetlights",
        "description": "Returns a single work order joined with the full streetlight objects (id, latitude, longitude, status, last_seen) for every light attached via work_order_lights. Use this to fetch everything an agent needs to render a work-order detail view — the parent work order plus the geo/status of each light it covers — in one round trip.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Work order ID",
            "schema": { "type": "integer", "example": 1 }
          }
        ],
        "responses": {
          "200": {
            "description": "Work order with full streetlight details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["id", "description", "status", "created_at", "lights"],
                  "properties": {
                    "id": { "type": "integer", "example": 1 },
                    "description": {
                      "type": "string",
                      "example": "Replace damaged fixture at Cedar Ln"
                    },
                    "status": {
                      "type": "string",
                      "enum": ["open", "closed"],
                      "example": "open"
                    },
                    "created_at": {
                      "type": "string",
                      "format": "date-time",
                      "example": "2024-01-15T09:00:00Z"
                    },
                    "lights": {
                      "type": "array",
                      "description": "Full streetlight rows for every light attached to this work order, ordered by light id. Empty array if no lights are attached.",
                      "items": {
                        "type": "object",
                        "required": ["id", "latitude", "longitude", "status", "last_seen"],
                        "properties": {
                          "id": { "type": "integer", "example": 3 },
                          "latitude": { "type": "number", "example": 30.26758 },
                          "longitude": { "type": "number", "example": -97.74556 },
                          "status": {
                            "type": "string",
                            "enum": ["ok", "fault", "assigned"],
                            "example": "assigned"
                          },
                          "last_seen": {
                            "type": "string",
                            "format": "date-time",
                            "example": "2024-01-15T08:30:00Z"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid id parameter (not an integer)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["error"],
                  "properties": {
                    "error": { "type": "string", "example": "id must be an integer" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["error"],
                  "properties": {
                    "error": { "type": "string", "example": "Unauthorized" }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Work order ID not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["error"],
                  "properties": {
                    "error": { "type": "string", "example": "Work order 99 not found" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/work-orders": {
      "get": {
        "operationId": "getWorkOrders",
        "summary": "List work orders",
        "description": "Returns all work orders, each with its assigned streetlight IDs.",
        "responses": {
          "200": {
            "description": "Array of work orders",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": ["id", "description", "status", "created_at", "light_ids"],
                    "properties": {
                      "id": { "type": "integer", "example": 1 },
                      "description": {
                        "type": "string",
                        "example": "Replace damaged fixture at Cedar Ln"
                      },
                      "status": {
                        "type": "string",
                        "enum": ["open", "closed"],
                        "example": "open"
                      },
                      "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-15T09:00:00Z"
                      },
                      "light_ids": {
                        "type": "array",
                        "items": { "type": "integer" },
                        "example": [3, 7]
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["error"],
                  "properties": {
                    "error": { "type": "string", "example": "Unauthorized" }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createWorkOrder",
        "summary": "Create one or more work orders",
        "description": "Creates work orders, links each to the specified streetlights, and atomically flips those lights' status to 'assigned'. Accepts either a single object to create one work order, or an array of objects to create many work orders in a single atomic transaction (all-or-nothing). All light IDs must exist; any missing ID causes a 400 with no partial writes. In batch (array) form, the same light_id appearing in two entries is rejected with a 400 (a light cannot be assigned to two work orders in the same call). Use the single-object form for a single work order; use the array form to create multiple work orders atomically.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "object",
                    "description": "Single work order — creates one work order.",
                    "required": ["description", "light_ids"],
                    "properties": {
                      "description": {
                        "type": "string",
                        "minLength": 1,
                        "description": "Non-empty description of the work to be done",
                        "example": "Inspect and repair lamp on Oak Blvd corridor"
                      },
                      "light_ids": {
                        "type": "array",
                        "items": { "type": "integer" },
                        "minItems": 1,
                        "description": "IDs of streetlights to assign to this work order",
                        "example": [3, 8, 9]
                      }
                    }
                  },
                  {
                    "type": "array",
                    "description": "Batch — creates all listed work orders in a single atomic transaction. Cross-entry duplicate light_ids are rejected.",
                    "minItems": 1,
                    "items": {
                      "type": "object",
                      "required": ["description", "light_ids"],
                      "properties": {
                        "description": {
                          "type": "string",
                          "minLength": 1,
                          "description": "Non-empty description of the work to be done",
                          "example": "Inspect and repair lamp on Oak Blvd corridor"
                        },
                        "light_ids": {
                          "type": "array",
                          "items": { "type": "integer" },
                          "minItems": 1,
                          "description": "IDs of streetlights to assign to this work order",
                          "example": [3, 8, 9]
                        }
                      }
                    }
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Work order(s) created successfully. Single-object request returns one work-order object; array request returns an array of work-order objects in the same order as the request.",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "object",
                      "description": "Returned for single-object requests.",
                      "required": ["id", "description", "status", "created_at", "light_ids", "lights"],
                      "properties": {
                        "id": { "type": "integer", "example": 3 },
                        "description": {
                          "type": "string",
                          "example": "Inspect and repair lamp on Oak Blvd corridor"
                        },
                        "status": {
                          "type": "string",
                          "enum": ["open", "closed"],
                          "example": "open"
                        },
                        "created_at": {
                          "type": "string",
                          "format": "date-time",
                          "example": "2024-01-15T10:00:00Z"
                        },
                        "light_ids": {
                          "type": "array",
                          "items": { "type": "integer" },
                          "example": [3, 8, 9]
                        },
                        "lights": {
                          "type": "array",
                          "description": "Full streetlight rows after status flip to 'assigned'",
                          "items": {
                            "type": "object",
                            "required": ["id", "latitude", "longitude", "status", "last_seen"],
                            "properties": {
                              "id": { "type": "integer", "example": 3 },
                              "latitude": { "type": "number", "example": 30.26758 },
                              "longitude": { "type": "number", "example": -97.74556 },
                              "status": {
                                "type": "string",
                                "enum": ["ok", "fault", "assigned"],
                                "example": "assigned"
                              },
                              "last_seen": {
                                "type": "string",
                                "format": "date-time",
                                "example": "2024-01-15T08:30:00Z"
                              }
                            }
                          }
                        }
                      }
                    },
                    {
                      "type": "array",
                      "description": "Returned for array requests — one entry per created work order, in request order.",
                      "items": {
                        "type": "object",
                        "required": ["id", "description", "status", "created_at", "light_ids", "lights"],
                        "properties": {
                          "id": { "type": "integer", "example": 3 },
                          "description": {
                            "type": "string",
                            "example": "Inspect and repair lamp on Oak Blvd corridor"
                          },
                          "status": {
                            "type": "string",
                            "enum": ["open", "closed"],
                            "example": "open"
                          },
                          "created_at": {
                            "type": "string",
                            "format": "date-time",
                            "example": "2024-01-15T10:00:00Z"
                          },
                          "light_ids": {
                            "type": "array",
                            "items": { "type": "integer" },
                            "example": [3, 8, 9]
                          },
                          "lights": {
                            "type": "array",
                            "description": "Full streetlight rows after status flip to 'assigned'",
                            "items": {
                              "type": "object",
                              "required": ["id", "latitude", "longitude", "status", "last_seen"],
                              "properties": {
                                "id": { "type": "integer", "example": 3 },
                                "latitude": { "type": "number", "example": 30.26758 },
                                "longitude": { "type": "number", "example": -97.74556 },
                                "status": {
                                  "type": "string",
                                  "enum": ["ok", "fault", "assigned"],
                                  "example": "assigned"
                                },
                                "last_seen": {
                                  "type": "string",
                                  "format": "date-time",
                                  "example": "2024-01-15T08:30:00Z"
                                }
                              }
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Validation error — bad body shape, empty array, bad description, bad light_ids, one or more light IDs not found, or (batch only) the same light_id appearing in two entries. In batch form, per-entry validation errors are prefixed with the offending index (e.g. 'entries[2]: description must be a non-empty string').",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["error"],
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Streetlight IDs not found: 99, 100"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["error"],
                  "properties": {
                    "error": { "type": "string", "example": "Unauthorized" }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}